Implements part of #51
1. Most functions on the cli/api.go file were helpers.
2. The root command doesn't need any function.
